Fix OPEN_BRACE check after function definition
There are two places where we check that the opening brace following
a function definition goes on the next line. The second one is more
extensive as it takes into account the context so it detects multiline
function arguments and return value on a separate line. So let's drop
the first check. The second check has a couple of flaws though: it
doesn't allow opening and closing braces on the same line and it doesn't
detect functions with attributes (e.g. inline). Fix them both.