Don't treat calls of functions from a scope as declarations
In C++ it's illegal to declare a function using a syntax involving the
scope access (`::`) operator. Nevertheless, the checkpatch identifies
the `class::method(a, &b)`, being a function call, as a declaration of
a function with nameless arguments.

The problem is that the checkpatch recognizes a sequence of identifiers
and preprocessor concatenation (`##`) with namespace access (`::`) as
a single identifier. For example: `class_name##_base::method`. If the
sequence only had a `##` operator, it could be considered as a single
identifier, but since it contains the `::` operator, it can't, so the
pattern detecting function declrations becomes invalid.

In order to not break other existing rules that could rely on the fact
the namespace access is a part of an identifier, this patch only fixes
the function declaration recognition by introducing a new pattern for
sequences of identifiers with preprocessor concatenation operators but
without namespace accesses and using it.