Allow passing empty last argument to macros
In C it's posssible to pass empty token sequences to a macro, for
example: `F(a, b,,)`, the third and fourth arguments of the F macro
invokation are empty token sequences.

Let's make the checkpatch understand having the last macro argument
empty (the `,)` sequence). This by the way allows `,]` and `,;` but
these are illegal in C anyways.