Fix use of undef value if no .gitattributes file
The `$generatedFiles` variable is uninitialized but compared with an
empty string if the `.gitattributes` file is missing. Let's initialize
the variable with an empty string, so missing `.gitattributes` file is
interpreted in the same way as no `linguist-generated=true` attributed
items in an existing `.gitattributes` file.

Fixes 2770f41d32c9 ("Fix bug when all files are skipped if
generatedFiles is empty").

Part of #76